6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,9 +1,11 @@
 #include <stdio.h>
+
 /**
  * main - this is the main function
  *
  * Return: 0 (Success)
  */
+
 int main(void)
 {
 	char a;
@@ -11,10 +13,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float e;
-        printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(a));
-	printf("Size of an int: %lu byte(s)\n",sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n",sizeof(c));
-	printf("Size of a long long int: %lu byte(s)\n",sizeof(d));
-	printf("Size of a float: %lu byte(s)\n",sizeof(e));
+        printf("Size of a char: %lu byte(s)\n",(unsigned long) sizeof(a));
+	printf("Size of an int: %lu byte(s)\n",(unsigned long) sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n",(unsigned long) sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n",(unsigned long) sizeof(d));
+	printf("Size of a float: %lu byte(s)\n",(unsigned long) sizeof(e));
         return (0);
 }
